Attempt #2 to fix UPSERT.

+++ sqlite/src/sqliteInt.h
@@ -2917,7 +2917,7 @@ struct Upsert {
   int iDataCur;             /* Index of the data cursor */
   int iIdxCur;              /* Index of the first index cursor */
 #if defined(SQLITE_BUILDING_FOR_COMDB2)
-  int oeFlag;               /* ON CONFLICT action flag */
+  u8 oeFlag;                /* ON CONFLICT action flag */
 #endif /* defined(SQLITE_BUILDING_FOR_COMDB2) */
 };
